create specific display fields file in helpers
