+++ packages/site/src/routes/[dictionaryId]/entry/[entryId]/seoDescription.ts
@@ -1,4 +1,10 @@
 import { orderGlosses } from '../../../../lib/helpers/glosses';
+import {
+  showLocalOrthographies,
+  showPartsOfSpeech,
+  removeItalicTagsWithAPeriod,
+  unnecessaryLineBreaksRegex,
+} from '../../../../lib/helpers/entry/displayFields';
 import type { IEntry } from '@living-dictionaries/types';
 
 // function previousMethod(entry: IEntry, dictionaryGlossLanguages: string[], $t: (key: string) => string) {
@@ -56,32 +62,3 @@ export function seoDescription(
 
   return description.replace(unnecessaryLineBreaksRegex, '').trim();
 }
-
-//TODO these might be helper functions instead, since we can use them in print view and in other places.
-export function showLocalOrthographies(entry: IEntry) {
-  const localOrthographiesRegex = /^(lo[2-5]*)$/gm; // exclusively matches: lo lo2 lo3 lo4 & lo5
-  const localOrthographiesKeys = Object.keys(entry).filter((entry) =>
-    entry.match(localOrthographiesRegex)
-  );
-  const localOrthographies = localOrthographiesKeys.map((lo) => entry[lo]);
-  return localOrthographies;
-}
-
-export function showPartsOfSpeech(pos: string | string[]) {
-  if (pos) {
-    if (typeof pos !== 'string' && pos.length > 1) {
-      return pos.join(', ') + '.';
-    }
-    return pos + '.';
-  }
-  return '';
-}
-
-export function removeItalicTagsWithAPeriod(str: string) {
-  if (str) {
-    return str.replace(/<\/?i>/g, '') + '.';
-  }
-  return '';
-}
-
-export const unnecessaryLineBreaksRegex = /(?<!\w)\n/gm;
